Allow stack to take a config parameter
Unless the stack method is explicitly overloaded by the class that is including
Restroom, method calls will fail because the connection method always passes it
a Faraday::Connection object. While it would be necessary to overload the method
for any APIs that require authentication, any open API will be unreachable.